Enable debug logging in Launcher and LauncherModel and the system property "debug.launcher.ignore-cache".
This is to help track down http://code.google.com/p/cyanogenmod/issues/detail?id=512,
and should be reverted if that bug is confirmed fixed.

If debug.launcher.ignore-cache is set, never load launcher resources out
of cache; always reload.  This is just to help diagnostics and will make
the launcher very slow; this should also be reverted if this issue is
fixed.