Fix possible race condition stopping mDesktopLoaderThread.
mRunning is set inside the thread, after it's already running.  It's
possible that the thread might be held up by other activity, and not
get to do anything for a while.  We could come back into loadUserItems
before mRunning is set; it wouldn't think the thread is running, so
it wouldn't try to stop it, leading to two DesktopLoaderThreads running.

Change "running" to "!finished", and consider the object not finished
as soon as it's created.  This requires that this object be only used
once, which it is.  It also requires that the object's thread be created
if it's created at all, or mFinished will never be set.

It would be a bit cleaner to set mRunning to true as soon as the Thread
is created, but there's no override in Runnable to do that; we'd need to
rearchitect this a bit to make DesktopItemsLoader itself be the thread,
rather than a runnable held by mDesktopLoaderThread.  It's not worth the
bigger change against upstream to do this.