Don't allow launcher thread joins to timeout.
This was the fundamental cause of launcher icons disappearing: ths code
blindly assumed that thread joins would never time out, and the exception
was simply thrown away, so the behavior when it did was never tested.

Some misbehaviors when this was happening have been fixed, but this really
shouldn't be allowed to happen.  If the thread isn't dying, then we really
should allow the FC dialog to pop up, and not just continue on and hope
everything works out.