Fix a launcher deadlock.
loadApplications and stopAndWaitForApplicationsLoader are synchronized,
and ApplicationsLoader::run synchronizes with LauncherModel, causing
a deadlock.

(The problem was exposed by removing the timeout from the thread joins.
Previously, it would deadlock for five seconds, then time out and march
on blindly.  This means that compared to CM5.0.7 before the disappearing
icons fix, this should fix some random 5-second stalls in the launcher.)

Remove the synchronize.  Even if mStopped is set to true after the conditional,
we still have, in fact, successfully loaded applications, so it's perfectly
valid to set mApplicationsLoaded.  The only caveat is that other threads can
only safely set mApplicationsLoaded back to false if the thread isn't running.

This means that abortLoaders shouldn't be setting mApplicationsLoaded = false,
because it doesn't end the thread; it just asynchronously asks the thread to
exit and then continues without waiting.  Leave it to the thread, when it
finally exits, to set mApplicationsLoaded correctly.

Note that loadApplications also sets mApplicationsLoaded = false.  That's correct
(though probably redundant); the thread is never running here (we just called
stopAndWaitForApplicationsLoader), and this is explicitly clearing the app cache
to reload it.

This has a side-effect: previously, abortLoaders had the unobvious side-effect
of clearing the app cache (clearing mApplicationsLoaded).  That no longer happens.
This function is only used by Launcher on shutdown anyway.